Adding take care of all tdsparser data sent directly to stream
