+++ src/System.Data.SqlClient/src/System/Data/SqlClient/SqlCommand.cs
@@ -1735,7 +1735,9 @@ namespace System.Data.SqlClient
             {
                 RegisterForConnectionCloseNotification(ref returnedTask);
 
-                Task<SqlDataReader>.Factory.FromAsync((commandBehavior, callback, stateObject) => BeginExecuteReader(callback, stateObject, commandBehavior), EndExecuteReader, behavior, null).ContinueWith((t) =>
+                Task<SqlDataReader>.Factory.FromAsync(
+                    (commandBehavior, callback, stateObject) => 
+                    BeginExecuteReader(callback, stateObject, commandBehavior), EndExecuteReader, behavior, null).ContinueWith((t) =>
                 {
                     registration.Dispose();
                     if (t.IsFaulted)
