+++ src/System.Data.SqlClient/src/System/Data/SqlClient/TdsParserStateObjectNative.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace System.Data.SqlClient
@@ -284,7 +285,7 @@ namespace System.Data.SqlClient
             }
         }
 
-        internal override void SetPacketData(PacketHandle packet, byte[] buffer, int bytesUsed)
+        private void SetPacketData(PacketHandle packet, byte[] buffer, int bytesUsed)
         {
             Debug.Assert(packet.Type == PacketHandle.NativePacketType, "unexpected packet type when requiring NativePacket");
             SNINativeMethodWrapper.SNIPacketSetData(packet.NativePacket, buffer, bytesUsed);
@@ -381,6 +382,269 @@ namespace System.Data.SqlClient
                     Clear();
                 }
             }
+        }
+
+
+        internal override Task WriteSni(bool canAccumulate)
+        {
+            // Prepare packet, and write to packet.
+            PacketHandle packet = GetResetWritePacket();
+
+            SetBufferSecureStrings();
+            SetPacketData(packet, _outBuff, _outBytesUsed);
+
+            uint sniError;
+            Debug.Assert(Parser.Connection._parserLock.ThreadMayHaveLock(), "Thread is writing without taking the connection lock");
+            Task task = SNIWritePacket(packet, out sniError, canAccumulate, callerHasConnectionLock: true);
+
+            // Check to see if the timeout has occurred.  This time out code is special case code to allow BCP writes to timeout. Eventually we should make all writes timeout.
+            if (_bulkCopyOpperationInProgress && 0 == GetTimeoutRemaining())
+            {
+                _parser.Connection.ThreadHasParserLockForClose = true;
+                try
+                {
+                    Debug.Assert(_parser.Connection != null, "SqlConnectionInternalTds handler can not be null at this point.");
+                    AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, (byte)0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
+                    _bulkCopyWriteTimeout = true;
+                    SendAttention();
+                    _parser.ProcessPendingAck(this);
+                    ThrowExceptionAndWarning();
+                }
+                finally
+                {
+                    _parser.Connection.ThreadHasParserLockForClose = false;
+                }
+            }
+
+            // Special case logic for encryption removal.
+            if (_parser.State == TdsParserState.OpenNotLoggedIn &&
+                _parser.EncryptionOptions == EncryptionOptions.LOGIN)
+            {
+                // If no error occurred, and we are Open but not logged in, and
+                // our encryptionOption state is login, remove the SSL Provider.
+                // We only need encrypt the very first packet of the login message to the server.
+
+                // We wanted to encrypt entire login channel, but there is
+                // currently no mechanism to communicate this.  Removing encryption post 1st packet
+                // is a hard-coded agreement between client and server.  We need some mechanism or
+                // common change to be able to make this change in a non-breaking fashion.
+                _parser.RemoveEncryption();                        // Remove the SSL Provider.
+                _parser.EncryptionOptions = EncryptionOptions.OFF; // Turn encryption off.
+
+                // Since this packet was associated with encryption, dispose and re-create.
+                ClearAllWritePackets();
+            }
+
+            SniWriteStatisticsAndTracing();
+
+            ResetBuffer();
+
+            AssertValidState();
+            return task;
+        }
+
+        // Sends an attention signal - executing thread will consume attn.
+        internal override void SendAttention(bool mustTakeWriteLock = false)
+        {
+            if (!_attentionSent)
+            {
+                // Dumps contents of buffer to OOB write (currently only used for
+                // attentions.  There is no body for this message
+                // Doesn't touch this._outBytesUsed
+                if (_parser.State == TdsParserState.Closed || _parser.State == TdsParserState.Broken)
+                {
+                    return;
+                }
+
+                PacketHandle attnPacket = CreateAndSetAttentionPacket();
+
+                try
+                {
+                    // Dev11 #344723: SqlClient stress hang System_Data!Tcp::ReadSync via a call to SqlDataReader::Close
+                    // Set _attentionSending to true before sending attention and reset after setting _attentionSent
+                    // This prevents a race condition between receiving the attention ACK and setting _attentionSent
+                    _attentionSending = true;
+
+#if DEBUG
+                    if (!_skipSendAttention)
+                    {
+#endif
+                        // Take lock and send attention
+                        bool releaseLock = false;
+                        if ((mustTakeWriteLock) && (!_parser.Connection.ThreadHasParserLockForClose))
+                        {
+                            releaseLock = true;
+                            _parser.Connection._parserLock.Wait(canReleaseFromAnyThread: false);
+                            _parser.Connection.ThreadHasParserLockForClose = true;
+                        }
+                        try
+                        {
+                            // Check again (just in case the connection was closed while we were waiting)
+                            if (_parser.State == TdsParserState.Closed || _parser.State == TdsParserState.Broken)
+                            {
+                                return;
+                            }
+
+                            uint sniError;
+                            _parser._asyncWrite = false; // stop async write 
+                            SNIWritePacket(attnPacket, out sniError, canAccumulate: false, callerHasConnectionLock: false);
+                        }
+                        finally
+                        {
+                            if (releaseLock)
+                            {
+                                _parser.Connection.ThreadHasParserLockForClose = false;
+                                _parser.Connection._parserLock.Release();
+                            }
+                        }
+#if DEBUG
+                    }
+#endif
+
+                    SetTimeoutSeconds(AttentionTimeoutSeconds); // Initialize new attention timeout of 5 seconds.
+                    _attentionSent = true;
+                }
+                finally
+                {
+                    _attentionSending = false;
+                }
+
+
+                AssertValidState();
+            }
+        }
+
+        internal Task SNIWritePacket(PacketHandle packet, out uint sniError, bool canAccumulate, bool callerHasConnectionLock)
+        {
+            // Check for a stored exception
+            var delayedException = Interlocked.Exchange(ref _delayedWriteAsyncCallbackException, null);
+            if (delayedException != null)
+            {
+                throw delayedException;
+            }
+
+            Task task = null;
+            _writeCompletionSource = null;
+
+            PacketHandle packetPointer = EmptyReadPacket;
+
+            bool sync = !_parser._asyncWrite;
+            if (sync && _asyncWriteCount > 0)
+            { // for example, SendAttention while there are writes pending
+                Task waitForWrites = WaitForAccumulatedWrites();
+                if (waitForWrites != null)
+                {
+                    try
+                    {
+                        waitForWrites.Wait();
+                    }
+                    catch (AggregateException ae)
+                    {
+                        throw ae.InnerException;
+                    }
+                }
+                Debug.Assert(_asyncWriteCount == 0, "All async write should be finished");
+            }
+            if (!sync)
+            {
+                // Add packet to the pending list (since the callback can happen any time after we call SNIWritePacket)
+                packetPointer = AddPacketToPendingList(packet);
+            }
+
+            // Async operation completion may be delayed (success pending).
+            try
+            {
+            }
+            finally
+            {
+                sniError = WritePacket(packet, sync);
+            }
+
+            if (sniError == TdsEnums.SNI_SUCCESS_IO_PENDING)
+            {
+                Debug.Assert(!sync, "Completion should be handled in SniManagedWwrapper");
+                Interlocked.Increment(ref _asyncWriteCount);
+                Debug.Assert(_asyncWriteCount >= 0);
+                if (!canAccumulate)
+                {
+                    // Create completion source (for callback to complete)
+                    _writeCompletionSource = new TaskCompletionSource<object>();
+                    task = _writeCompletionSource.Task;
+
+                    // Ensure that setting _writeCompletionSource completes before checking _delayedWriteAsyncCallbackException
+                    Interlocked.MemoryBarrier();
+
+                    // Check for a stored exception
+                    delayedException = Interlocked.Exchange(ref _delayedWriteAsyncCallbackException, null);
+                    if (delayedException != null)
+                    {
+                        throw delayedException;
+                    }
+
+                    // If there are no outstanding writes, see if we can shortcut and return null
+                    if ((_asyncWriteCount == 0) && ((!task.IsCompleted) || (task.Exception == null)))
+                    {
+                        task = null;
+                    }
+                }
+            }
+#if DEBUG
+            else if (!sync && !canAccumulate && SqlCommand.DebugForceAsyncWriteDelay > 0)
+            {
+                // Executed synchronously - callback will not be called 
+                TaskCompletionSource<object> completion = new TaskCompletionSource<object>();
+                uint error = sniError;
+                new Timer(obj =>
+                {
+                    try
+                    {
+                        if (_parser.MARSOn)
+                        { // Only take reset lock on MARS.
+                            CheckSetResetConnectionState(error, CallbackType.Write);
+                        }
+
+                        if (error != TdsEnums.SNI_SUCCESS)
+                        {
+                            AddError(_parser.ProcessSNIError(this));
+                            ThrowExceptionAndWarning();
+                        }
+                        AssertValidState();
+                        completion.SetResult(null);
+                    }
+                    catch (Exception e)
+                    {
+                        completion.SetException(e);
+                    }
+                }, null, SqlCommand.DebugForceAsyncWriteDelay, Timeout.Infinite);
+                task = completion.Task;
+            }
+#endif
+            else
+            {
+                if (_parser.MARSOn)
+                { // Only take reset lock on MARS.
+                    CheckSetResetConnectionState(sniError, CallbackType.Write);
+                }
+
+                if (sniError == TdsEnums.SNI_SUCCESS)
+                {
+                    _lastSuccessfulIOTimer._value = DateTime.UtcNow.Ticks;
+
+                    if (!sync)
+                    {
+                        // Since there will be no callback, remove the packet from the pending list
+                        Debug.Assert(IsValidPacket(packetPointer), "Packet added to list has an invalid pointer, can not remove from pending list");
+                        RemovePacketFromPendingList(packetPointer);
+                    }
+                }
+                else
+                {
+                    AddError(_parser.ProcessSNIError(this));
+                    ThrowExceptionAndWarning(callerHasConnectionLock);
+                }
+                AssertValidState();
+            }
+            return task;
         }
     }
 }
